+++ apps/next-app-zod/openapi-gen.config.ts
@@ -119,7 +119,7 @@ export default defineConfig({
     outputFile: "arazzo.yaml",
   },
   overlay: {
-    version: "1.1.0",
+    version: "1.2.0",
     apply: ["./overlays/public.overlay.yaml"],
   },
 });